Move prototype flask app into multinet directory
