+++ .env
@@ -2,5 +2,5 @@ EXTERNAL_MONGO_PORT=27021
 GIRDER_SERVE_PORT=9090
 DOC_SERVE_PORT=8000
 
-FLASK_APP=flaskinet
+FLASK_APP=multinet
 FLASK_ENV=development
