--- flaskinet/db.py
@@ -1,1 +1,0 @@
-../multinet/multinet/db.py
