--- flaskinet/multinet.gql
@@ -1,1 +1,0 @@
-../multinet/multinet/multinet.gql
