--- flaskinet/resolvers
@@ -1,1 +1,0 @@
-../multinet/multinet/resolvers
