--- flaskinet/types.py
@@ -1,1 +1,0 @@
-../multinet/multinet/types.py
